pipeline-modules/scripts/pipeline/requirements.txt
rich>=13.0.0
pyyaml>=6.0
prompt-toolkit>=3.0.0
